rename gc -> globus compute
