+++ src/appfl/algorithm/__init__.py
@@ -3,7 +3,7 @@ This module provides a base class to define the functions required to implement 
 """
 from .client_optimizer import *
 from .client_step_optimizer import *
-from .gc_client_optimizer import *
+from .globus_compute_client_optimizer import *
 from .personalized_client_optimizer import *
 from .server_fed_avg import *
 from .server_fed_avgmom import *
